requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot[job-queue]==20.9
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
python-dotenv==1.0.1
cloudinary==1.41.0
httpx==0.27.2
apscheduler==3.10.4
